backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "eventlet>=0.40.3",
    "fastmcp>=2.12.3",
    "flask>=3.1.2",
    "flask-socketio>=5.5.1",
    "gunicorn>=23.0.0",
    "openai>=1.108.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-socketio>=5.13.0",
]
